+++ src/Modules/FUN_MODULE/Death.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Nadybot\Modules\FUN_MODULE;
+
+use Nadybot\Core\DBRow;
+
+class Death extends DBRow {
+	public string $character;
+	public int $counter = 0;
+}